Chapter_001/src/test/java/ru/job4j/it/ListUtilsTest.java
package ru.job4j.it;

import org.hamcrest.core.Is;
import org.junit.Test;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.*;

public class ListUtilsTest {

    @Test
    public void whenAddBefore() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 3));
        ListUtils.addBefore(input, 1, 2);
        assertThat(Arrays.asList(1, 2, 3), Is.is(input));
    }

    @Test(expected = IndexOutOfBoundsException.class)
    public void whenAddBeforeWithInvalidIndex() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 3));
        ListUtils.addBefore(input, 3, 2);
    }

    @Test
    public void whenAddAfter() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 3));
        ListUtils.addAfter(input, 0, 2);
        assertThat(Arrays.asList(1, 2, 3), Is.is(input));
    }

    @Test
    public void whenRemoveIfMoreThen3() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5));
        List<Integer> result = ListUtils.removeIf(input, x -> x > 3);
        assertThat(result, Is.is(Arrays.asList(1, 2, 3)));
    }

    @Test
    public void whenReplaceIfMoreThen3() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5));
        List<Integer> result = ListUtils.replaceIf(input, x -> x > 3, 0);
        assertThat(result, Is.is(Arrays.asList(1, 2, 3, 0, 0)));
    }

    @Test
    public void whenRemoveAll() {
        List<Integer> input = new ArrayList<>(Arrays.asList(1, 2, 3, 4, 5, 6));
        List<Integer> forRemove = Arrays.asList(4, 5, 6);
        List<Integer> result = ListUtils.removeAll(input, forRemove);
        assertThat(result, Is.is(Arrays.asList(1, 2, 3)));
    }

}